Dockerize with conda
