+++ extractors-mrcnn/requirements.txt
@@ -10,3 +10,4 @@ opencv-python
 h5py
 imgaug
 IPython[all]
+pycocotools==2.0.1
